requirements.txt
beautifulsoup4==4.4.1
dj-database-url==0.4.0
Django==1.9.2
django-nocaptcha-recaptcha==0.0.18
django-simple-captcha==0.5.1
django-simple-math-captcha==1.0.7
djangorestframework==3.3.3
gunicorn==19.4.5
oauthlib==1.0.3
Pillow==3.1.1
psycopg2==2.6.1
PyJWT==1.4.0
python-openid==2.2.5
python-social-auth==0.2.14
requests==2.9.1
requests-oauthlib==0.6.1
six==1.10.0
virtualenv==15.0.0
whitenoise==2.0.6